+++ src/Smartstore.Core/Platform/Identity/Domain/Customer.cs
@@ -343,7 +343,6 @@ namespace Smartstore.Core.Identity
         /// <summary>
         /// Gets or sets the wallet history.
         /// </summary>
-        [IgnoreDataMember]
         public ICollection<WalletHistory> WalletHistory
         {
             get => _walletHistory ?? LazyLoader.Load(this, ref _walletHistory) ?? (_walletHistory ??= new HashSet<WalletHistory>());