+++ src/Smartstore.Modules/Smartstore.Stripe/wwwroot/smartstore.stripe.js
@@ -31,24 +31,21 @@
     }
 
     return {
-        initPaymentElement: function (publicApiKey, secret, apiVersion) {
-            stripe = Stripe(publicApiKey, {
-                apiVersion: apiVersion,
-                betas: ['elements_enable_deferred_intent_beta_1'],
-            });
+        initPaymentElement: function (publicApiKey, apiVersion, amount, currency, captureMethod) {
+            stripe = Stripe(publicApiKey, { apiVersion: apiVersion });
 
-            const { clientSecret } = { clientSecret: secret };
-
-            const appearance = {
-                theme: 'stripe',
+            const options = {
+                mode: 'payment',
+                amount: amount,
+                currency: currency,
+                captureMethod: captureMethod,
+                appearance: { theme: 'stripe' },
+                paymentMethodCreation: "manual"
             };
 
-            elements = stripe.elements({ appearance, clientSecret });
+            elements = stripe.elements(options);
 
-            const paymentElementOptions = {
-                layout: "tabs",
-            };
-
+            const paymentElementOptions = { layout: "tabs" };
             const paymentElement = elements.create("payment", paymentElementOptions);
             paymentElement.mount(paymentElementSelector);
 
@@ -65,11 +62,7 @@
 
             // Listen for changes to the radio input elements.
             $(document, "input[name='paymentmethod']").on("change", function (e) {
-                if (e.target.value == moduleSystemName) {
-                    btnNext[0].disabled = true;
-                } else {
-                    btnNext[0].disabled = false;
-                }
+                btnNext[0].disabled = e.target.value == moduleSystemName;
             });
 
             // Handle button state on page load
@@ -83,10 +76,17 @@
             $("form").on("submit", async e => {
                 if ($("input[name='paymentmethod']:checked").val() == moduleSystemName && !createdPaymentMethod) {
                     e.preventDefault();
+
+                    // Trigger form validation and wallet collection
+                    const { error: submitError } = await elements.submit();
+                    if (submitError) {
+                        displayNotification(submitError.message, 'error');
+                        return;
+                    }
+
                     (async () => {
-                        const { error, paymentMethod } = await stripe.createPaymentMethod({
-                            elements
-                        });
+                        const { error, paymentMethod } = await stripe.createPaymentMethod({ elements });
+
                         $.ajax({
                             type: 'POST',
                             data: {