+++ src/Smartstore.Modules/Smartstore.WebApi/Controllers/Identity/CustomersController.cs
@@ -64,6 +64,17 @@ namespace Smartstore.Web.Api.Controllers
             return GetRelatedQuery(key, x => x.Orders);
         }
 
+        /// <summary>
+        /// Gets WalletHistory entities assigned to a Customer.
+        /// </summary>
+        /// <remarks>Only applicable if a wallet plugin is installed.</remarks>
+        [HttpGet("Customers({key})/WalletHistory"), ApiQueryable]
+        [Permission("Wallet.read")]
+        public IQueryable<WalletHistory> GetWalletHistory(int key)
+        {
+            return GetRelatedQuery(key, x => x.WalletHistory);
+        }
+
         [HttpGet("Customers({key})/ReturnRequests"), ApiQueryable]
         [Permission(Permissions.Order.ReturnRequest.Read)]
         public IQueryable<ReturnRequest> GetReturnRequests(int key)