+++ src/Smartstore.Web/Components/OrderTotalsViewComponent.cs
@@ -100,7 +100,7 @@ namespace Smartstore.Web.Components
                     && isEditable;
             }
 
-            if (isEditable && (_shoppingCartSettings.AllowToDeactivateCartItems ||
+            if (isEditable && (_shoppingCartSettings.AllowActivatableCartItems ||
                 await _shoppingCartService.CountProductsInCartAsync(customer, ShoppingCartType.ShoppingCart, storeId, false) > 0))
             {
                 model.SubtotalLabel = T("ShoppingCart.Totals.SubTotalSelectedProducts", model.TotalQuantity);